fix(oil detection) model accounts for the text in simulation
The old model cropped the text from the simulation from INPE, which caused it to detect the text as oil.
Now the masks igore the text but still retain them in the grayscale images, this way the model learns to ignore the text.
Also added instructions on how to compile the code to a binary file.